Bugfix: Fail when /etc/smlmp.conf exists but is unreadable
Previous version simply checks for existence of /etc/smlmp.conf, and if
it exists, uses config.read on it, which will silently fail if the file
can't be read but exists, i.e. with wrong permissions. Since config.read
returns a list of files that it successfully read, we can check if it
returned something; if not, the config file wasn't read, and we enter a
fail condition.